#119 Todo検索強化(コードの共通化)
